resolved cyclic import dependencies by merging savefile classes into one file
