+++ index.ts
@@ -1,6 +1,4 @@
 export * from "./models/savefile.ts";
-export * from "./models/srmsavefile.ts";
-export * from "./models/srasavefile.ts";
 export * from "./models/saveheader.ts";
 export * from "./models/saveslot.ts";
 export * from "./models/scene.ts";
